[WIP] Split vertex fuser into a separate file
